mercurial: add a config knob to apply patches to the `tip` revision
Add a `use_latest_revision` config knob to the repo configuration that
causes patch stacks to be applied to the latest revision in the repo
(`tip`) instead of the parent revision of the base commit in the stack.
This is useful for uplift revisions, where the parent of the base
may not exist in the repo, and where applying the patch to the latest
is desirable to get early feedback on any merge conflicts that will be
encountered on landing.

Move the logic for determining the initial `base_revision` into a
separate function to enable a simple unit test.